\name{BayesFactorNGS}
\alias{BayesFactorNGS}
\docType{data}
\title{ The Binary Negative Gold Standard Developed via Bayes Factor}
\description{
  This data consists of the 7739 binary interactions with computed
  Bayes factor of less than -3. This dataset was created 16 June 2009.
}
\usage{data(BayesFactorNGS)}
\format{
  A data frame with  observations on the following 5 variables.
  \describe{
    \item{\code{Orf1}}{The ORF for gene 1}
    \item{\code{Orf2}}{The ORF for gene 2}
    \item{\code{Tested}}{The number of times the interaction was tested}
    \item{\code{Observed}}{The number of times the interaction was found}
    \item{\code{log_BF}}{The log of the resulting Bayes factor}
    
  }
}
\details{
  None.
}
\source{
 }
\references{
 }
\examples{
data(BayesFactorNGS)
}
\keyword{datasets}
