\name{BinaryGS}
\alias{BinaryGS}
\docType{data}
\title{ The Binary Gold Standard Data set Reported by Yu et al}
\description{
  This data consists of the 1318 binary interactions Yu et al reported
  as their binary gold standard data set.
}
\usage{data(BinaryGS)}
\format{
  A data frame with 1318 observations on the following 2 variables.
  \describe{
    \item{\code{ORF1}}{The ORF for gene 1}
    \item{\code{ORF2}}{The ORF for gene 2}
  }
}
\details{
  None.
}
\source{
  The data were downloaded from
  \url{http://interactome.dfci.harvard.edu/S_cerevisiae} on Nov 21, 2008.
}
\references{
 H. Yu et al, High Quality Binary Protein Interaction Map of the Yeast
 Interactome Network, Science 322, 104, 2008.
}
\examples{
data(BinaryGS)
}
\keyword{datasets}
